remove old comments, discussion relation in own file
