+++ pkg/models/plan_discussion.go
@@ -1,9 +1,5 @@
 package models
 
-import (
-	"github.com/google/uuid"
-)
-
 //PlanDiscussion represents a discussion that a user has about a model plan
 type PlanDiscussion struct {
 	BaseStruct
@@ -30,18 +26,3 @@ const (
 	DiscussionWaiting    DiscussionStatus = "WAITING_FOR_RESPONSE"
 	DiscussionUnAnswered DiscussionStatus = "UNANSWERED"
 )
-
-//IDiscussionRelation is an interface that represents models that are related to a discussion.
-type IDiscussionRelation interface {
-	GetDiscussionID() uuid.UUID
-}
-
-//DiscussionRelation is an embedded struct meant to satisify the IDiscussionRelation interface
-type DiscussionRelation struct {
-	DiscussionID uuid.UUID `json:"discussionID" db:"discussion_id"`
-}
-
-//GetDiscussionID returns DiscussionID
-func (d DiscussionRelation) GetDiscussionID() uuid.UUID {
-	return d.DiscussionID
-}
