+++ pkg/models/plan_document.go
@@ -8,7 +8,6 @@ import (
 type PlanDocument struct {
 	BaseStruct
 	ModelPlanRelation
-	// ModelPlanID uuid.UUID `json:"modelPlanID" db:"model_plan_id"`
 
 	FileType string `json:"fileType" db:"file_type"`
 	Bucket   string `json:"bucket" db:"bucket"`
